Refactor tests to Django testing
